src/use-cases/authenticate-org.use-case.ts
import { OrgsRepository } from '@/repositories/orgs-repository';
import type { ORG } from '@prisma/client';
import { compare } from 'bcryptjs';
import { InvalidCredentialsError } from './errors/invalid-credentials.error';

interface AuthenticateOrgUseCaseRequest {
  email: string;
  password: string;
}

interface AuthenticateOrgUseCaseResponse {
  org: ORG;
}

export class AuthenticateOrgUseCase {
  constructor(private orgsRepository: OrgsRepository) {}

  async execute({
    email,
    password,
  }: AuthenticateOrgUseCaseRequest): Promise<AuthenticateOrgUseCaseResponse> {
    const org = await this.orgsRepository.findByEmail(email);

    if (!org) {
      throw new InvalidCredentialsError();
    }

    const doestPasswordMatches = await compare(password, org.password);

    if (!doestPasswordMatches) {
      throw new InvalidCredentialsError();
    }

    return { org };
  }
}
